// components/knockout-table/types.tsx
import { MatchProps } from "@/types/match";

export interface KnockoutTableProps {
  isNational?: boolean;
  roundOf16_1: MatchProps[] | [];
  roundOf16_2: MatchProps[] | [];
  roundOf16_3: MatchProps[] | [];
  roundOf16_4: MatchProps[] | [];
  roundOf16_5: MatchProps[] | [];
  roundOf16_6: MatchProps[] | [];
  roundOf16_7: MatchProps[] | [];
  roundOf16_8: MatchProps[] | [];
  quarterFinals_1: MatchProps[] | [];
  quarterFinals_2: MatchProps[] | [];
  quarterFinals_3: MatchProps[] | [];
  quarterFinals_4: MatchProps[] | [];
  semiFinals_1: MatchProps[] | [];
  semiFinals_2: MatchProps[] | [];
  final: MatchProps[] | [];
}